backend/requirements.txt
Flask==3.1.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Werkzeug==3.1.0
SQLAlchemy==2.0.45
